Ensure atomicity of updates
- save the manifest into a temporary file, and then move it into place.
  Otherwise it would be observed as empty during our write.

- delete unreferenced files *after* the manifest is written, so if we
  failed to write the manifest we don't leave things in an unusable
  state.

There's a remaining problem here, which is that a reader who looks at an
old manifest and then reads referenced files can be interrupted by a
fetch which deletes some of those files (having written an updated manifest).
Probably the solution there is to be more reticent to delete old files:
eg, deleting unreferenced files which are older than 30 days.  That still
fulfills the goal of avoiding unlimited storage use.